Running for Windows
Seems to have issues when it comes to building. I also went from electron-rebuild to electron-builder (I think the building is the same). Seems that anything above electron-webpack 2.3.1 doesnt work well so I might specify version to strict electron-webpack 2.3.1. I hope to keep it to ^2.3.1 to keep everything nice.

Not receiving any data back unlike when I did it on macOS