Modified structure to support production build
Other things that I did are fix some files to follow the new eslint rules.

Main thing is that electron-webpack has issues rendering files that are not explicitly stated. In a technical more sense, anytime `fs.readdirSync` is called, the production build (turned to an executable file) does not work due to this.

Also modified electron-webpack to support adding environment variables (written in .env) to the environment on the production build.