Added unit testing to project. Uses mocha/chai/spectron to allow for unit testing.
Make sure you run npm install again as there are several new packages required to run these tests.

Also added a very simple (stub) for the vehicle class just to allow for the tests to run & be somewhat meaningful.
This class can be entirely replaced with the actual code when that class gets written.