Formatting update
Removed "_" from functions that started with a "_" per request. Map.js didn't have the "_" so I decided to keep it the same with other react components.

Renamed "test" to "fixtures." Will create a real test folder where we perform unit tests on the react components.

Will be testing the program's ability to run on a build later on this week.