Multi-window support for electron-webpack
Allows us to create more than one window on electron and linking them all to one html file without collision.

This works for building the app too, not just on development.

Next step is to update the new mission window to display whatever information is necessary for mission selected. See #45