Updated CSS structure
Note: when importing files (import X from './X'), and there are two files X.js and x.css, the program will import the CSS file (when intended to import the JS file).

This will cause an error in the program. The solution is to not name the CSS file as JS files.

Took a while to figure out since all code looked normal. Ill get working on getting structure up (will do Takeoff as an example)