Small change
Current version struggles to keep mission window state as well as job state.

The mission window can only keep track of one misison at a time, but it should be able to keep track of all missions. If I start one mission and start another, the mission window will forget all information about the first mission. There should be a fix to this.