Added basic file to ListDict (now called DictionaryList)
DictionaryList makes more sense than ListDict, since this is a dictionary of lists, not a list of dictionaries.

Also edited/fixed comments to help with the `atom-typescript` plugin when hovering over variables.

Re-allocated declaration files.

Turned on a check that will crash the program before loading if MAPBOX_TOKEN is not defined.