Reformatted configuration files and types
Changed exports in static/index so that the variables can be seen as configuration files when imported to other files. Will cause less confusion to the developer.

Moved all source types to a `src/types` folder. Note this is different from external module typings, which will contain types for external modules.

Commented out things that do not work right now because of these changes, and will fix them back once I have the time.

Added typings to all tasks and messages.

TODO:

Fix code that I have commented with code that works for it now.

Finish the type guard functions in src/types/messages. @tnleang can you help me with this?

Fix documentation for all of this. Perhaps add a `docs` folder in root?