Updated type guards
Fixed one small thing with connect message as well as removed the requirement for an errorMessage when a vehicle is in an error state.

Also removed the requirement for an error field in badMessage message, but we will be filling that field in whenever we send a badMessage. This is in the case a vehicle does not provide an error field in the badMessage message.