Added mission parameters
Mission parameters is a mix of information required, as well as options.

A selected number of missions can be chosen to perform at once, but since the output of one mission goes into the input of the next, the user only needs to add mission info for the first mission. The user will need to input mission options for all missions, however.