Added more logic to Mission
Went on with Max's definition of handleUnresponsiveVehicle.

Added support to add tasks to a job by a certain comparison criteria (allows for obtaining the next task in a different way).

Ensured that activeVehicleMapping ONLY has the vehicles that are doing a job that relates to the mission. Any other vehicles are completely ignored for the mission (this is in the case the user interface sends extra jobs that are irrelevant to the mission over to the constructor). Doing this allows us to not have to check all the time if the job in the activeVehicleMapping is a related jobType or not. It simply is.

Added support to keep track of vehicles of a mission by having a callback on the update messages that are sent from the vehicles.