Fixed logic + comments
Added check to Orchestrator to check Mission's vehicles before calling mission.update()

Did check on all vehicle states before assigning jobs in Mission (instead of checking states while assigning jobs).

Added check for new message for bad messages too.

Fixed logic in Orchestrator to set completionParameters of this mission to the starting parameters for next mission.

Logic in Mission will be to send stop to all vehicles if the stop() function is called while the Mission status is not "ready". If Mission status is "initializing" it will also stop sending current messages.

Added logic to Vehicle's status handler to remove the callback when it disconnects, as the other callback should not happen if the vehicle is too late to accept its job.