Modified fixtures
Updated logic to Orchestrator. UpdateVehiclesFixtures now sends notifications to Orchestrator instead of the container.

Fixed bug with above where Orchestrator does not initialize the vehicle to "ready" status, and denies all messages since it is in a "disconnected" status.

TODO: Add logic to send stop message if vehicle is in a mission and sends status = "running" or status = "paused".

Moved acknowledging to MessageHandler. No need for it to be in Orchestrator, was thinking that MessageHandler needed vehicle access to send acknowledgement, but that's not needed.