Update logic in Orchestrator's ping + ack/self change
Moved logic that requires only new message in an if loop, and put the Orchestrator acknowledge outside of that.

Simplified loop of creating new event handler every few seconds with delta timeout to a simple timeout. Prevented disconnectFromVehicle to happening when vehicle is already disconnected (due to recursive calls after the program ends).

Fixed logic in vehicle to not create a new instance every time it is reconnected.

Added logic in the Orchestrator to send "stop" message to vehicle if it is a mission-related state (waiting, running, paused) while it is not assigned a mission.

Fixed logic in MessageHandler as msgpack.decode actually obtains the object itself, not the string. The object can vary from undefined to a valid message. Also fixed the jobTypes type guard (see src/static/index)