Added logic to MissionWindow
Will change the logic in MissionWindow as the other apps become built. The way the other apps will work is that the apps will not send information to the MissionWindow unless all their fields are filled out.

Other changes:

Changed how the bounding box looks if it is locked (the draggable markers will be gone)

Changed commenting to the fix in leaflet's default marker in WaypointMarker.

Fixed layout choosing in App.tsx (not using any type anymore).