Added pause/resume logic + cache information
Cache information is to store the information when user puts information to the Mission UI but the user changes window afterwards. We do not want the information to be lost and make the user reinput the information. Not sure if having a cache will work, but if it doesnt, will remove.