Logic to Orchestrator/Mission + etc
Big change to Orchestrator to provide an actual function callback, instead of passing the function, as passing the function will NOT update the instance variables of the Orchestrator. Big note for the future.

Initialize Mission after constructing it from the Ochestrator, as its abstract values do not apply until the constructor is done. Before, the mission name would never be applied since the initialize would automatically be called from the Orchestrator.

Simplified updateVehicle fixtures to only emit "ready" status to help test missions.

Other fixes to Mission class and subclasses that were buggy. Will continue to fix these issues tonight