Added logic to support POI makers in VTOL Search mission
Not tested. Fixes checkmark 2 of #68.

Logic is to keep track of the detailed search task that is assigned, and see if a POI message is sent from the VTOL (that is doing the detailed search job). If so, that POI is valid. If a complete message is sent with no POI message in between, it means that the POI is invalid.