Update package dependencies and pass tests
 ## 1) Why is the change being made?

Before this change, all the packages of GCS are outdated and tests for
current commits are failing.

 ## 2) What has changed to address the problem?

This commit updates the packages that GCS uses and ensures that all
tests pass and that all scripts run properly. The `npm-check-updates`
package was used to determine which packages needed to be updated.

There are a few packages that were not updated:

(1) electron has stayed at `electron@3` due to the fact any future
electron updates does not let the application run (usually due to
serialport not binding properly). A future to do will
be to update this electron version and actually make it run.

(2) rc-slider has stayed at `rc-slider@8` due to the fact
@types/rc-slider does not work with `rc-slider@9`

 ## 3) How was this change tested?

This change was tested with the following:

(1) `npm test` to check for linting errors, TypeScript errors, and unit
test errors (no changes to unit testing).

(2) `npm start` to check if the application runs properly and the
`electron` version is compatible with the `electron-webpack` version as well
as `serialport`.

(3) `npm build` and running the executable to ensure that all modules
build properly with `electron-webpack` and `electron-builder`.