Simplify eslint with airbnb-typescript module
 ## Why is the change being made?

This change is made because I ran across the
`eslint-plugin-airbnb-typescript` plugin, which puts the
`@typescript-eslint` and `eslint-plugin-airbnb` package together well.
Before, this package was not available as the `@typescript-eslint`
package just came out.

 ## What has changed to address the problem?

This change modifies `eslintrc.yml` config file to remove many rules
and simplifies it with the `airbnb-typescript` plugin. The only rule
kept is the rule to not throw eslint errors to chai unit test assertions.

 ## 3) How was this change tested?

This change was tested with `npm test`.