Force build to fail with test fail
This change addresses the fact that Travis builds actually do not fail
when certain scripts fail (because the script exits with a 0/success)
but the errors have not been emitted yet.

This change changes any `&` to `&&` in `package.json` so it is run in
sequence (will result in an error if any fails instead of quiet error).
`travis.yml` is also changed to run `npm test` instead of running the
individual scripts.