Add rules to make GCS repo support LF better
 ## Why is the change being made?

This change is made because .rst files are ending with CRLF, when all
files in the project should end in LF.

 ## What has changed to address the problem?

This change will modify `.gitattributes` to make .rst files end with
LF. `.gitattributes` also removed inclusion of .c and .h files since
this is not a C project.

This change also adds documentation to setting up git properly so that
non-Unix systems (mainly Windows) will not put CRLF to files when they
commit.

 ## How was this change tested?

This change was tested with `npm test`.