Merge docs script into one npm script
 ## Why is the change being made?

This change is made because the scripts for running docs (one was for
Windows and the other for macOS) was redundant.

 ## What has changed to address the problem?

This change merges the scripts into one script. `npm run start:windocs`
and `npm run start:macdocs` is now merged to `npm run docs` and should
work.

 ## How was this change tested?

This change was tested with `npm run docs` on Windows. This has not been
tested on macOS or Linux, but I am confident that it works too.