Refactor and simplify documentation
This change increases legibility of raw documentation files to make it
easier for newer developers to document.

The following changes are specifics to what have happened with format:

- Explicity show where borders are by adding more `------` to
the existing border.

- Separate links to have the link and target definition.
Before a link would show as `test <somelink.com>`_. Now it is `test`_
and the link is in the bottom of the page.

- Make any link that is left together to be anonymous to stop erros from
coming from `npm run docs` (seen in the console but is not reported when
testing). Before a link that would show as `test2 <anotherlink.com>`_
would be `test2 <anotherlink.com>`__

- Make `note` annotations inline.

The following are some bigger refactoring of files:

- Deleted `ncu.rst`. The project now uses dependabot to manage
dependencies.

- Deleted `rules.rst`. The rules are very basic. This page will have
more specific rules in the future.

- Moved `toctree` (table of content tree) of Mission Examples to the
bottom of the page in `examples.rst`.
- any standard link
- move example tree in `examples.rst` to bottom of page
- compressed `note` in packages to be inline