Migrate docs to Markdown and update docs
 ## Why is the change being made?

This change is made because reStructured text could be a more difficult
style of documentation to adapt to (for newer developers) than Markdown.

 ## What has changed to address the problem?

This change removes the Sphinx reStructured text documentation in GCS
and replaces it with a MkDocs Markdown text documentation structure.

Everyone contributing to the project will have to follow the
instructions under `docs/contributing/getting-started.md`.

There is additional added documentation to address the following:
- How to work with GCS (running, building, and testing)
- Installing MkDocs to see docs locally in `localhost:8000`.
- Adding control messages for manual vehicle control
- Adding geofencing field to start messages

 ## How was this change tested?

This change was tested by checking if the docs looked fine when running
`npm run docs`.