Add coverage through istanbul + coveralls
 ## Why is the change being made?

This change is made because GCS has no coverage currently. It is
important to have coverage in code to see how much code is actually
unit tested.

 ## What has changed to address the problem?

This change adds coverage to the code. The coverage right now is pretty
bad but it can always be improved in the future.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits