Update Travis CI to add webhooks and remove cache
 ## Why is the change being made?

This change is made because webhooks from GitHub to Discord was annoying
with pushes/pull requests/etc and was not necessary. The necessary
update to developers is the build status of the `master` branch.

This change also addresses the fact that a build has failed to succeed
due to an invalid cache.

 ## What has changed to address the problem?

This change will add a script that will run another script that sends a
webhook of the build. The webhook URL has to be set up in Discord, and
made an environmental variable in Travis CI to work.

 ## How was this change tested?

This change will be tested by making the `webhook` branch emit to
Discord and see if it goes through successfully. Then the emitting
branch will be changed to `master`.

Also, the cache was removed from `.travis.yml`.

 ## Related documents, URLs, commits

https://github.com/DiscordHooks/travis-ci-discord-webhook#guide