Allow GCS to use latest version of electron
 ## Why is the change being made?

This change is made because GCS was running on electron v3, a version
made back two years ago. This version is long deprecated and GCS needed
to upgrade to v8 sooner or later. However, electron v5 would not build
at all (we dont consider v4 as it does not support geolocation).

 ## What has changed to address the problem?

This change finds the issue to be the fact electron v5+ turns off the
`nodeIntegration` field for windows (used to be on by default). This was
breaking our application.

This change sets `nodeIntegration` to true for all windows created, and
updates electron to v8. Other related fields that were modified from v3
to v8 are updated in this change.

Other minor changes are the following:
- removing dependency on `electron-rebuild` and setting `preinstall` script
to `electron-builder install-app-deps` in `package.json`
- removing `build` field in `package.json` as `electron-rebuild` is no
longer a dependency
- added dependency to `@types/node` v12 as default version v13 breaks tests with
electron v8 (see https://github.com/electron/electron/issues/21612)

 ## How was this change tested?

This change was tested with `npm test`, running the application and
seeing if it was fine, and running the application from `npm run build`
and seeing if it was fine (checking console, build status, and if
serialport runs).

 ## Related documents, URLs, commits

https://github.com/electron/electron/releases/tag/v5.0.0