Add script to test build in Travis
 ## Why is the change being made?

This change is made because it is important for Travis to see if the
build works by running the build script and seeing if it passes.

 ## What has changed to address the problem?

This change adds `npm run build` to Travis's test scripts.

 ## How was this change tested?

This test will be tested through Travis. A GH_TOKEN environment variable
will be provided to Travis for the build to pass.

 ## Related documents, URLs, commits