Add GeofenceMessage to gcs-vehicles-messages.md
Why is the change being made?

This change is made because GCS needs a new message to handle Geofencing
 information outside of the start message. This change is
 not necessary, but it improves scalability for future use of GCS.

What has changed to address the problem?

This change created a new GeofenceMessage containing geofencing values
that can be sent to all vehicles. StartMessage was changed as well by
 deleting the geofencing parameters that was in it previously.

How was this change tested?
This change was tested using npm test.