Add KeepIn and KeepOut geofence boxes
  ## Why is this being changed?

This is being changed because of Northrop's requirement of having a
KeepIn and KeepOut box for all vehicles. The KeepIn area is an area that
all vehicles must stay inside of, and will be about the size of the
mission field/runway. The KeepOut area is an area that all vehicles must
stay outside of, and will be the location of all student/Northrop
observers.

  ## What was changed to address this problem?

In order to address the problem, two geofence buttons were created
inside of the missionWindow UI tah tallows the GCS operator to
dynamically create the two boxes individually, separating them by
color.

  ## How was this change tested?

This change was tested with npm test.