Add support to save console log as a .log file
 ## Why is the change being made?

This change is made to address issue #53 as it is needed for improvement
of the application.

 ## What has changed to address the problem?

In order to address the problem, a Save Log button was created next to
the Clear Log button in the LogContainer. This button calls saveMessages
which preprocess the log messages as a string then pass it to
ipc.postSaveLog. Then postSaveLog pass the messages to saveLog function
in index.ts through ipcRenderer.

 ## How was this change tested?

This change was tested with npm test.